requirements.txt
httpx
feedparser
google-genai
networkx
sqlalchemy
fastapi
uvicorn[standard]
apscheduler
tenacity
python-dotenv
